[MOVE] user repository 관련 클래스 user 패키지 하위로 이동
